+++ bw.workflow.editconvictedreport/src/main/java/sa/gov/nic/bio/bw/workflow/editconvictedreport/controllers/EditJudgmentDetailsPaneFxController.java
@@ -1,14 +1,17 @@
 package sa.gov.nic.bio.bw.workflow.editconvictedreport.controllers;
 
+import javafx.beans.binding.Bindings;
 import javafx.beans.binding.BooleanBinding;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.value.ChangeListener;
 import javafx.fxml.FXML;
+import javafx.geometry.NodeOrientation;
 import javafx.scene.Node;
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
+import javafx.scene.control.Label;
 import javafx.scene.control.RadioButton;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.Pane;
@@ -39,21 +42,41 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 {
 	@Input(alwaysRequired = true) private List<CrimeCode> oldCrimes;
 	@Input(alwaysRequired = true) private JudgementInfo judgementInfo;
-	@Output private String judgmentIssuer;
-	@Output private String judgmentNumber;
-	@Output private LocalDate judgmentDate;
+	@Output private String judgmentIssuerOldValue;
+	@Output private String judgmentIssuerNewValue;
+	@Output private String judgmentNumberOldValue;
+	@Output private String judgmentNumberNewValue;
+	@Output private LocalDate judgmentDateOldValue;
+	@Output private LocalDate judgmentDateNewValue;
 	@Output private Boolean judgmentDateUseHijri;
-	@Output private String caseFileNumber;
-	@Output private Long prisonerNumber;
-	@Output private LocalDate arrestDate;
+	@Output private String caseFileNumberOldValue;
+	@Output private String caseFileNumberNewValue;
+	@Output private Long prisonerNumberOldValue;
+	@Output private Long prisonerNumberNewValue;
+	@Output private LocalDate arrestDateOldValue;
+	@Output private LocalDate arrestDateNewValue;
 	@Output private Boolean arrestDateUseHijri;
 	@Output private List<CrimeCode> newCrimes;
 	
+	@FXML private Pane paneJudgmentDetails;
 	@FXML private Pane paneCrimeContainer;
 	@FXML private Pane paneCrime2;
 	@FXML private Pane paneCrime3;
 	@FXML private Pane paneCrime4;
 	@FXML private Pane paneCrime5;
+	@FXML private Pane paneJudgmentIssuerReset;
+	@FXML private Pane paneJudgmentNumberReset;
+	@FXML private Pane paneJudgmentDateReset;
+	@FXML private Pane paneCaseFileNumberReset;
+	@FXML private Pane panePrisonerNumberReset;
+	@FXML private Pane paneArrestDateReset;
+	@FXML private Pane paneCrimeDiffContainer;
+	@FXML private Pane paneOldCrimes;
+	@FXML private Pane paneNewCrimes;
+	@FXML private TextField txtJudgmentIssuer;
+	@FXML private TextField txtJudgmentNumber;
+	@FXML private TextField txtCaseFileNumber;
+	@FXML private TextField txtPrisonerNumber;
 	@FXML private ComboBox<ComboBoxItem<Integer>> cboCrimeEvent1;
 	@FXML private ComboBox<ComboBoxItem<Integer>> cboCrimeClass1;
 	@FXML private ComboBox<ComboBoxItem<Integer>> cboCrimeEvent2;
@@ -64,20 +87,47 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 	@FXML private ComboBox<ComboBoxItem<Integer>> cboCrimeClass4;
 	@FXML private ComboBox<ComboBoxItem<Integer>> cboCrimeEvent5;
 	@FXML private ComboBox<ComboBoxItem<Integer>> cboCrimeClass5;
-	@FXML private TextField txtJudgmentIssuer;
-	@FXML private TextField txtJudgmentNumber;
-	@FXML private TextField txtCaseFileNumber;
-	@FXML private TextField txtPrisonerNumber;
-	@FXML private DatePicker dpArrestDate;
-	@FXML private DatePicker dpJudgmentDate;
 	@FXML private RadioButton rdoArrestDateUseHijri;
 	@FXML private RadioButton rdoArrestDateUseGregorian;
 	@FXML private RadioButton rdoJudgmentDateUseHijri;
 	@FXML private RadioButton rdoJudgmentDateUseGregorian;
+	@FXML private DatePicker dpArrestDate;
+	@FXML private DatePicker dpJudgmentDate;
+	@FXML private Label lblJudgmentIssuerOldValue;
+	@FXML private Label lblJudgmentIssuerNewValue;
+	@FXML private Label lblJudgmentNumberOldValue;
+	@FXML private Label lblJudgmentNumberNewValue;
+	@FXML private Label lblJudgmentDateOldValue;
+	@FXML private Label lblJudgmentDateNewValue;
+	@FXML private Label lblCaseFileNumberOldValue;
+	@FXML private Label lblCaseFileNumberNewValue;
+	@FXML private Label lblPrisonerNumberOldValue;
+	@FXML private Label lblPrisonerNumberNewValue;
+	@FXML private Label lblArrestDateOldValue;
+	@FXML private Label lblArrestDateNewValue;
+	@FXML private Label lblCrime1NewValue;
+	@FXML private Label lblCrime2NewValue;
+	@FXML private Label lblCrime3NewValue;
+	@FXML private Label lblCrime4NewValue;
+	@FXML private Label lblCrime5NewValue;
+	@FXML private Glyph iconJudgmentIssuerArrow;
+	@FXML private Glyph iconJudgmentNumberArrow;
+	@FXML private Glyph iconJudgmentDateArrow;
+	@FXML private Glyph iconCaseFileNumberArrow;
+	@FXML private Glyph iconPrisonerNumberArrow;
+	@FXML private Glyph iconArrestDateArrow;
+	@FXML private Glyph iconCrimeDiff;
+	@FXML private Button btnJudgmentIssuerReset;
+	@FXML private Button btnJudgmentNumberReset;
+	@FXML private Button btnJudgmentDateReset;
+	@FXML private Button btnCaseFileNumberReset;
+	@FXML private Button btnPrisonerNumberReset;
+	@FXML private Button btnArrestDateReset;
 	@FXML private Button btnHidePaneCrime2;
 	@FXML private Button btnHidePaneCrime3;
 	@FXML private Button btnHidePaneCrime4;
 	@FXML private Button btnHidePaneCrime5;
+	@FXML private Button btnCrimesReset;
 	@FXML private Button btnAddMore;
 	@FXML private Button btnPrevious;
 	@FXML private Button btnStartOver;
@@ -274,6 +324,15 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 		initCrimeEventComboBox(cboCrimeEvent4, cboCrimeClass4, crimeEventTitles, crimeClassTitles);
 		initCrimeEventComboBox(cboCrimeEvent5, cboCrimeClass5, crimeEventTitles, crimeClassTitles);
 		
+		judgmentIssuerOldValue = judgementInfo.getJudgIssuer();
+		judgmentNumberOldValue = judgementInfo.getJudgNum();
+		Long judgmentDateLong = judgementInfo.getJudgDate();
+		if(judgmentDateLong != null) judgmentDateOldValue = AppUtils.secondsToGregorianDate(judgmentDateLong);
+		caseFileNumberOldValue = judgementInfo.getPoliceFileNum();
+		prisonerNumberOldValue = judgementInfo.getPrisonerNumber();
+		Long arrestDateLong = judgementInfo.getArrestDate();
+		if(arrestDateLong != null) arrestDateOldValue = AppUtils.secondsToGregorianDate(arrestDateLong);
+		
 		Node focusedNode = null;
 		String judgmentIssuer = null;
 		String judgmentNumber = null;
@@ -285,50 +344,204 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 		
 		if(isFirstLoad())
 		{
-			judgmentIssuer = judgementInfo.getJudgIssuer();
-			judgmentNumber = judgementInfo.getJudgNum();
-			//judgmentDate = judgementInfo.getJudgDate();
-			caseFileNumber = judgementInfo.getPoliceFileNum();
-			prisonerNumber = judgementInfo.getPrisonerNumber();
-			//arrestDate = judgementInfo.getArrestDate();
+			judgmentIssuer = judgmentIssuerOldValue;
+			judgmentNumber = judgmentNumberOldValue;
+			judgmentDate = judgmentDateOldValue;
+			caseFileNumber = caseFileNumberOldValue;
+			prisonerNumber = prisonerNumberOldValue;
+			arrestDate = arrestDateOldValue;
+			crimes = oldCrimes;
 		}
 		
-		// TODO: I stopped here. I will continue after vacation.
-		
-		if(judgmentIssuer != null && !judgmentIssuer.isEmpty()) txtJudgmentIssuer.setText(judgmentIssuer);
-		else if(this.judgmentIssuer != null && !this.judgmentIssuer.isEmpty())
-		{
-			txtJudgmentIssuer.setText(this.judgmentIssuer);
-		}
+		if(judgmentIssuer != null) txtJudgmentIssuer.setText(judgmentIssuer);
+		else if(this.judgmentIssuerNewValue != null) txtJudgmentIssuer.setText(this.judgmentIssuerNewValue);
 		else focusedNode = txtJudgmentIssuer;
 		
-		if(judgmentNumber != null && !judgmentNumber.isEmpty()) txtJudgmentNumber.setText(judgmentNumber);
-		else if(this.judgmentNumber != null && !this.judgmentNumber.isEmpty())
-		{
-			txtJudgmentNumber.setText(this.judgmentNumber);
-		}
+		if(judgmentNumber != null) txtJudgmentNumber.setText(judgmentNumber);
+		else if(this.judgmentNumberNewValue != null) txtJudgmentNumber.setText(this.judgmentNumberNewValue);
 		else if(focusedNode == null) focusedNode = txtJudgmentNumber;
 		
 		if(judgmentDate != null) dpJudgmentDate.setValue(judgmentDate);
-		else if(this.judgmentDate != null) dpJudgmentDate.setValue(this.judgmentDate);
+		else if(this.judgmentDateNewValue != null) dpJudgmentDate.setValue(this.judgmentDateNewValue);
 		else if(focusedNode == null) focusedNode = dpJudgmentDate;
 		
 		rdoJudgmentDateUseHijri.setSelected(true);
 		if(this.judgmentDateUseHijri != null && !this.judgmentDateUseHijri)
 																		rdoJudgmentDateUseGregorian.setSelected(true);
 		
-		if(caseFileNumber != null && !caseFileNumber.isEmpty()) txtCaseFileNumber.setText(caseFileNumber);
+		if(caseFileNumber != null) txtCaseFileNumber.setText(caseFileNumber);
+		else if(this.caseFileNumberNewValue != null) txtCaseFileNumber.setText(this.caseFileNumberNewValue);
+		
 		if(prisonerNumber != null) txtPrisonerNumber.setText(String.valueOf(prisonerNumber));
+		else if(this.prisonerNumberNewValue != null) txtPrisonerNumber.setText(String.valueOf(
+																						this.prisonerNumberNewValue));
+		
 		if(arrestDate != null) dpArrestDate.setValue(arrestDate);
+		else if(this.arrestDateNewValue != null) dpArrestDate.setValue(this.arrestDateNewValue);
 		
 		rdoArrestDateUseHijri.setSelected(true);
 		if(this.arrestDateUseHijri != null && !this.arrestDateUseHijri) rdoArrestDateUseGregorian.setSelected(true);
 		
-		if(crimes != null)
+		if(crimes == null) crimes = newCrimes;
+		
+		for(int i = 0; i < crimes.size(); i++)
 		{
-			for(int i = 0; i < crimes.size(); i++)
+			CrimeCode crimeCode = crimes.get(i);
+			
+			if(crimeCode != null)
 			{
-				CrimeCode crimeCode = crimes.get(i);
+				if(i > 0) GuiUtils.showNode(cboCrimePaneMap.get(i), true);
+				
+				cboCrimeEventMap.get(i).getItems()
+						.stream()
+						.filter(item -> item.getItem().equals(crimeCode.getCrimeEvent()))
+						.findFirst()
+						.ifPresent(cboCrimeEventMap.get(i)::setValue);
+				cboCrimeClassMap.get(i).getItems()
+						.stream()
+						.filter(item -> item.getItem().equals(crimeCode.getCrimeClass()))
+						.findFirst()
+						.ifPresent(cboCrimeClassMap.get(i)::setValue);
+			}
+		}
+		
+		if(newCrimes == null)
+		{
+			newCrimes = new ArrayList<>();
+			updateNewCrimes();
+		}
+		
+		checkForCrimeDuplicates.run();
+		
+		if(focusedNode != null) focusedNode.requestFocus();
+		else btnNext.requestFocus();
+		
+		lblCrime1NewValue.textProperty().bind(Bindings.createStringBinding(() -> cboCrimeEvent1.getValue().getText() +
+																		" - " + cboCrimeClass1.getValue().getText(),
+                                                    cboCrimeEvent1.valueProperty(), cboCrimeClass1.valueProperty()));
+		lblCrime2NewValue.textProperty().bind(Bindings.createStringBinding(() -> cboCrimeEvent2.getValue().getText() +
+	                                                                    " - " + cboCrimeClass2.getValue().getText(),
+                                                    cboCrimeEvent2.valueProperty(), cboCrimeClass2.valueProperty()));
+		lblCrime3NewValue.textProperty().bind(Bindings.createStringBinding(() -> cboCrimeEvent3.getValue().getText() +
+	                                                                    " - " + cboCrimeClass3.getValue().getText(),
+                                                    cboCrimeEvent3.valueProperty(), cboCrimeClass3.valueProperty()));
+		lblCrime4NewValue.textProperty().bind(Bindings.createStringBinding(() -> cboCrimeEvent4.getValue().getText() +
+	                                                                    " - " + cboCrimeClass4.getValue().getText(),
+                                                    cboCrimeEvent4.valueProperty(), cboCrimeClass4.valueProperty()));
+		lblCrime5NewValue.textProperty().bind(Bindings.createStringBinding(() -> cboCrimeEvent5.getValue().getText() +
+	                                                                    " - " + cboCrimeClass5.getValue().getText(),
+                                                    cboCrimeEvent5.valueProperty(), cboCrimeClass5.valueProperty()));
+		
+		lblCrime2NewValue.visibleProperty().bind(paneCrime2.visibleProperty());
+		lblCrime2NewValue.managedProperty().bind(paneCrime2.managedProperty());
+		lblCrime3NewValue.visibleProperty().bind(paneCrime3.visibleProperty());
+		lblCrime3NewValue.managedProperty().bind(paneCrime3.managedProperty());
+		lblCrime4NewValue.visibleProperty().bind(paneCrime4.visibleProperty());
+		lblCrime4NewValue.managedProperty().bind(paneCrime4.managedProperty());
+		lblCrime5NewValue.visibleProperty().bind(paneCrime5.visibleProperty());
+		lblCrime5NewValue.managedProperty().bind(paneCrime5.managedProperty());
+		
+		boolean rtl = Context.getGuiLanguage().getNodeOrientation() == NodeOrientation.RIGHT_TO_LEFT;
+		FontAwesome.Glyph arrowIcon = rtl ? FontAwesome.Glyph.LONG_ARROW_LEFT : FontAwesome.Glyph.LONG_ARROW_RIGHT;
+		iconJudgmentIssuerArrow.setIcon(arrowIcon);
+		iconJudgmentNumberArrow.setIcon(arrowIcon);
+		iconJudgmentDateArrow.setIcon(arrowIcon);
+		iconCaseFileNumberArrow.setIcon(arrowIcon);
+		iconPrisonerNumberArrow.setIcon(arrowIcon);
+		iconArrestDateArrow.setIcon(arrowIcon);
+		iconCrimeDiff.setIcon(arrowIcon);
+		
+		paneJudgmentIssuerReset.visibleProperty().bind(
+				Bindings.notEqual(lblJudgmentIssuerOldValue.textProperty(), lblJudgmentIssuerNewValue.textProperty()));
+		paneJudgmentNumberReset.visibleProperty().bind(
+				Bindings.notEqual(lblJudgmentNumberOldValue.textProperty(), lblJudgmentNumberNewValue.textProperty()));
+		paneJudgmentDateReset.visibleProperty().bind(
+				Bindings.notEqual(lblJudgmentDateOldValue.textProperty(), lblJudgmentDateNewValue.textProperty()));
+		paneCaseFileNumberReset.visibleProperty().bind(
+				Bindings.notEqual(lblCaseFileNumberOldValue.textProperty(), lblCaseFileNumberNewValue.textProperty()));
+		panePrisonerNumberReset.visibleProperty().bind(
+				Bindings.notEqual(lblPrisonerNumberOldValue.textProperty(), lblPrisonerNumberNewValue.textProperty()));
+		paneArrestDateReset.visibleProperty().bind(
+				Bindings.notEqual(lblArrestDateOldValue.textProperty(), lblArrestDateNewValue.textProperty()));
+		
+		BooleanBinding crimesChangingBinding = Bindings.createBooleanBinding(() ->
+		{
+			updateNewCrimes();
+			return !sameCrimes(oldCrimes, newCrimes);
+		}, cboCrimeClass1.valueProperty(), cboCrimeClass2.valueProperty(), cboCrimeClass3.valueProperty(),
+           cboCrimeClass4.valueProperty(), cboCrimeClass5.valueProperty(), cboCrimeEvent1.valueProperty(),
+           cboCrimeEvent2.valueProperty(), cboCrimeEvent3.valueProperty(), cboCrimeEvent4.valueProperty(),
+           cboCrimeEvent5.valueProperty(), btnAddMore.armedProperty(), btnHidePaneCrime2.armedProperty(),
+           btnHidePaneCrime3.armedProperty(), btnHidePaneCrime4.armedProperty(),
+           btnHidePaneCrime5.armedProperty(), btnCrimesReset.armedProperty());
+		
+		paneCrimeDiffContainer.visibleProperty().bind(crimesChangingBinding);
+		
+		lblJudgmentIssuerOldValue.setText(judgementInfo.getJudgIssuer() != null ? judgementInfo.getJudgIssuer() : "");
+		lblJudgmentNumberOldValue.setText(judgementInfo.getJudgNum() != null ? judgementInfo.getJudgNum() : "");
+		lblJudgmentDateOldValue.setText(judgementInfo.getJudgDate() != null ?
+                                GuiUtils.formatLocalDate(AppUtils.secondsToGregorianDate(judgementInfo.getJudgDate()),
+                                                         rdoJudgmentDateUseHijri.isSelected()) : "");
+		lblCaseFileNumberOldValue.setText(judgementInfo.getPoliceFileNum() != null ?
+		                                  judgementInfo.getPoliceFileNum() : "");
+		lblPrisonerNumberOldValue.setText(judgementInfo.getPrisonerNumber() != null ?
+				                          String.valueOf(judgementInfo.getPrisonerNumber()) : "");
+		lblArrestDateOldValue.setText(judgementInfo.getArrestDate() != null ?
+                                GuiUtils.formatLocalDate(AppUtils.secondsToGregorianDate(judgementInfo.getArrestDate()),
+                                                         rdoArrestDateUseHijri.isSelected()) : "");
+		
+		for(CrimeCode oldCrime : oldCrimes)
+		{
+			String crimeEvent = crimeEventTitles.get(oldCrime.getCrimeEvent());
+			String crimeClass = crimeClassTitles.get(oldCrime.getCrimeClass());
+			Label label = new Label(crimeEvent + " - " + crimeClass);
+			label.getStyleClass().add("old-value");
+			paneOldCrimes.getChildren().add(label);
+		}
+		
+		btnJudgmentIssuerReset.setOnAction(actionEvent ->
+		{
+		    txtJudgmentIssuer.setText(judgementInfo.getJudgIssuer() != null ? judgementInfo.getJudgIssuer() : "");
+			paneJudgmentDetails.requestFocus();
+		});
+		btnJudgmentNumberReset.setOnAction(actionEvent ->
+		{
+			txtJudgmentNumber.setText(judgementInfo.getJudgNum() != null ? judgementInfo.getJudgNum() : "");
+			paneJudgmentDetails.requestFocus();
+		});
+		btnJudgmentDateReset.setOnAction(actionEvent ->
+		{
+			dpJudgmentDate.setValue(judgementInfo.getJudgDate() != null ?
+			                        AppUtils.secondsToGregorianDate(judgementInfo.getJudgDate()) : null);
+			paneJudgmentDetails.requestFocus();
+		});
+		btnCaseFileNumberReset.setOnAction(actionEvent ->
+		{
+			txtCaseFileNumber.setText(judgementInfo.getPoliceFileNum() != null ? judgementInfo.getPoliceFileNum() : "");
+			paneJudgmentDetails.requestFocus();
+		});
+		btnPrisonerNumberReset.setOnAction(actionEvent ->
+		{
+			txtPrisonerNumber.setText(judgementInfo.getPrisonerNumber() != null ?
+			                          String.valueOf(judgementInfo.getPrisonerNumber()) : "");
+			paneJudgmentDetails.requestFocus();
+		});
+		btnArrestDateReset.setOnAction(actionEvent ->
+		{
+			dpArrestDate.setValue(judgementInfo.getArrestDate() != null ?
+		                          AppUtils.secondsToGregorianDate(judgementInfo.getArrestDate()) : null);
+			paneJudgmentDetails.requestFocus();
+		});
+		btnCrimesReset.setOnAction(actionEvent ->
+		{
+			GuiUtils.showNode(paneCrime2, false);
+			GuiUtils.showNode(paneCrime3, false);
+			GuiUtils.showNode(paneCrime4, false);
+			GuiUtils.showNode(paneCrime5, false);
+			
+			for(int i = 0; i < oldCrimes.size(); i++)
+			{
+				CrimeCode crimeCode = oldCrimes.get(i);
 				
 				if(crimeCode != null)
 				{
@@ -346,12 +559,22 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 							.ifPresent(cboCrimeClassMap.get(i)::setValue);
 				}
 			}
-		}
+			paneJudgmentDetails.requestFocus();
+		});
 		
-		checkForCrimeDuplicates.run();
+		rdoJudgmentDateUseHijri.selectedProperty().addListener((observable, oldValue, newValue) ->
+                                lblJudgmentDateOldValue.setText(judgementInfo.getJudgDate() != null ?
+                                GuiUtils.formatLocalDate(AppUtils.secondsToGregorianDate(judgementInfo.getJudgDate()),
+                                                                         rdoJudgmentDateUseHijri.isSelected()) : ""));
+		rdoArrestDateUseHijri.selectedProperty().addListener((observable, oldValue, newValue) ->
+                               lblArrestDateOldValue.setText(judgementInfo.getArrestDate() != null ?
+						       GuiUtils.formatLocalDate(AppUtils.secondsToGregorianDate(judgementInfo.getArrestDate()),
+                                                                            rdoArrestDateUseHijri.isSelected()) : ""));
 		
-		if(focusedNode != null) focusedNode.requestFocus();
-		else btnNext.requestFocus();
+		if(paneCrime2.isVisible()) visibleCrimeTypesCount++;
+		if(paneCrime3.isVisible()) visibleCrimeTypesCount++;
+		if(paneCrime4.isVisible()) visibleCrimeTypesCount++;
+		if(paneCrime5.isVisible()) visibleCrimeTypesCount++;
 	}
 	
 	@Override
@@ -364,38 +587,26 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 	public void onGoingNext(Map<String, Object> uiDataMap)
 	{
 		var judgmentIssuer = txtJudgmentIssuer.getText();
-		if(!judgmentIssuer.isBlank()) this.judgmentIssuer = judgmentIssuer;
-		else this.judgmentIssuer = null;
+		if(!judgmentIssuer.isBlank()) this.judgmentIssuerNewValue = judgmentIssuer;
+		else this.judgmentIssuerNewValue = null;
 		
 		var judgmentNumber = txtJudgmentNumber.getText();
-		if(!judgmentNumber.isBlank()) this.judgmentNumber = judgmentNumber;
-		else this.judgmentNumber = null;
+		if(!judgmentNumber.isBlank()) this.judgmentNumberNewValue = judgmentNumber;
+		else this.judgmentNumberNewValue = null;
 		
-		this.judgmentDate = dpJudgmentDate.getValue();
+		this.judgmentDateNewValue = dpJudgmentDate.getValue();
 		this.judgmentDateUseHijri = rdoJudgmentDateUseHijri.isSelected();
 		
 		var caseFileNumber = txtCaseFileNumber.getText();
-		if(!caseFileNumber.isBlank()) this.caseFileNumber = caseFileNumber;
-		else this.caseFileNumber = null;
+		if(!caseFileNumber.isBlank()) this.caseFileNumberNewValue = caseFileNumber;
+		else this.caseFileNumberNewValue = null;
 		
 		var prisonerNumber = txtPrisonerNumber.getText();
-		if(!prisonerNumber.isBlank()) this.prisonerNumber = Long.parseLong(prisonerNumber);
-		else this.prisonerNumber = null;
+		if(!prisonerNumber.isBlank()) this.prisonerNumberNewValue = Long.parseLong(prisonerNumber);
+		else this.prisonerNumberNewValue = null;
 		
-		this.arrestDate = dpArrestDate.getValue();
+		this.arrestDateNewValue = dpArrestDate.getValue();
 		this.arrestDateUseHijri = rdoArrestDateUseHijri.isSelected();
-		
-		//crimes = new ArrayList<>();
-		//crimes.add(new CrimeCode(cboCrimeEvent1.getValue().getItem(), cboCrimeClass1.getValue().getItem()));
-		//
-		//for(int i = 1; i <= 4; i++)
-		//{
-		//	if(cboCrimePaneMap.get(i).isVisible())
-		//	{
-		//		crimes.add(new CrimeCode(cboCrimeEventMap.get(i).getValue().getItem(),
-		//		                         cboCrimeClassMap.get(i).getValue().getItem()));
-		//	}
-		//}
 	}
 	
 	private void initCrimeEventComboBox(ComboBox<ComboBoxItem<Integer>> cboCrimeEvent,
@@ -422,5 +633,42 @@ public class EditJudgmentDetailsPaneFxController extends WizardStepFxControllerB
 		
 		});
 		cboCrimeEvent.getSelectionModel().selectFirst();
+	}
+	
+	private void updateNewCrimes()
+	{
+		newCrimes.clear();
+		newCrimes.add(new CrimeCode(cboCrimeEvent1.getValue().getItem(), cboCrimeClass1.getValue().getItem()));
+		
+		for(int i = 1; i <= 4; i++)
+		{
+			if(cboCrimePaneMap.get(i).isVisible())
+			{
+				newCrimes.add(new CrimeCode(cboCrimeEventMap.get(i).getValue().getItem(),
+				                            cboCrimeClassMap.get(i).getValue().getItem()));
+			}
+		}
+	}
+	
+	private static boolean sameCrimes(List<CrimeCode> crimeCodes1, List<CrimeCode> crimeCodes2)
+	{
+		if(crimeCodes1 == null || crimeCodes2 == null) return false;
+		if(crimeCodes1.size() != crimeCodes2.size()) return false;
+		
+		outer: for(int i = 0; i < crimeCodes1.size(); i++)
+		{
+			for(int j = 0; j < crimeCodes1.size(); j++)
+			{
+				CrimeCode crimeCode1 = crimeCodes1.get(i);
+				CrimeCode crimeCode2 = crimeCodes2.get(j);
+				
+				if(crimeCode1.getCrimeEvent() == crimeCode2.getCrimeEvent() &&
+						crimeCode1.getCrimeClass() == crimeCode2.getCrimeClass()) continue outer;
+			}
+			
+			return false;
+		}
+		
+		return true;
 	}
 }