+++ RELEASE-NOTES.md
@@ -11,6 +11,17 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ---
 
+## v2019.04.1
+
+- ENH: Unify the transition overlay for menu transition and wizard transition.
+- ENH: Upgrade controlsfx library from 9.0.0 to 11.0.0.
+- NEW: Add Edit-Convicted-Report workflow.
+- CHN: Add 2 more options as fingerprints source: byCivilBiometricsId and byCriminalBiometricsId.
+- CHN: Add more criteria in search-convicted-report-by-search-criteria workflow.
+- CHN: Add history of all previous copies in convicted-report-dialog.
+
+---
+
 ## v2019.03.5
 
 - FIX: Fix Delete-Complete-Criminal-Record workflow logic.